pyproject.toml
[project]
name = "jasmine"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "dm-pix>=0.4.3",
  "einops>=0.8.0",
  "flax==0.11.2",
  "jax[cuda]==0.7.2",
  "optax>=0.2.3",
  "tyro>=0.8.5",
  "wandb[media]>=0.17.4",
  "grain==0.2.10",
  "array-record>=0.7.2",
  "pre-commit>=4.2.0",
  "pillow>=11.3.0",
  "jaxlpips",
  "imageio[ffmpeg]>=2.37.2",
  "minerl",
  "decord>=0.6.0",
  "torch>=2.1.0",
  "torchvision",
  "torchaudio",
  "diffusers==0.34.0",
  "huggingface-hub>=0.27.0",
  "robosuite==1.4.1",
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
minerl = { git = "https://github.com/minerllabs/minerl" }
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }
torchaudio = { index = "pytorch-cu121" }

[tool.uv]
override-dependencies = [
    "numpy>=1.26.0",
    "huggingface-hub>=0.27.0,<1.0",
    "diffusers>=0.34.0",
]
